feat: let the main thread spawn compute thread and render thread

+++ src/maze/cell.rs
@@ -20,7 +20,7 @@ impl GridCell {
     pub const START: GridCell = GridCell::Path(PathType::Start);
     pub const VISITED: GridCell = GridCell::Path(PathType::Visited);
     /// The width of each cell when rendered, in character widths.
-    pub const CELL_WIDTH: usize = 2;
+    pub const CELL_WIDTH: u16 = 2;
 }
 
 /// Represents different types of path cells in the maze.
@@ -62,7 +62,7 @@ impl fmt::Display for GridCell {
         };
         assert_eq!(
             styled_symbol.content().width(),
-            GridCell::CELL_WIDTH,
+            GridCell::CELL_WIDTH as usize,
             "Each cell must occupy exactly two character widths."
         );
         write!(f, "{}", styled_symbol)
